requirements.txt
asgiref==3.6.0
Django==4.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-environ==0.11.2
django-filter==23.1
django-js-asset==2.0.0
django-taggit==3.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
Pillow==9.5.0
psycopg2-binary==2.9.9
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
whitenoise==6.6.0
